Fixed implicit casts

+++ Inc/stm_utils.h
@@ -10,24 +10,24 @@
 
 #include <stdint.h>
 
-#define BIT0 0x0001
-#define BIT1 0x0002
-#define BIT2 0x0004
-#define BIT3 0x0008
-#define BIT4 0x0010
-#define BIT5 0x0020
-#define BIT6 0x0040
-#define BIT7 0x0080
-#define BIT8 0x0100
-#define BIT9 0x0200
-#define BITA 0x0400
-#define BITB 0x0800
-#define BITC 0x1000
-#define BITD 0x2000
-#define BITE 0x4000
-#define BITF 0x8000
+#define BIT0 ((uint32_t)0x0001)
+#define BIT1 ((uint32_t)0x0002)
+#define BIT2 ((uint32_t)0x0004)
+#define BIT3 ((uint32_t)0x0008)
+#define BIT4 ((uint32_t)0x0010)
+#define BIT5 ((uint32_t)0x0020)
+#define BIT6 ((uint32_t)0x0040)
+#define BIT7 ((uint32_t)0x0080)
+#define BIT8 ((uint32_t)0x0100)
+#define BIT9 ((uint32_t)0x0200)
+#define BITA ((uint32_t)0x0400)
+#define BITB ((uint32_t)0x0800)
+#define BITC ((uint32_t)0x1000)
+#define BITD ((uint32_t)0x2000)
+#define BITE ((uint32_t)0x4000)
+#define BITF ((uint32_t)0x8000)
 
-#define UPPER16BITS(BITFIELD) (BITFIELD << 16)
+#define UPPER16BITS(BITFIELD) (((uint32_t)BITFIELD) << 16)
 
 #define SET_BIT(BITFIELD, N) (BITFIELD |= ((uint32_t)0x1 << N))
 #define CLEAR_BIT(BITFIELD, N) (BITFIELD &= ~((uint32_t)0x1 << N))
